add global variables
